[HCB] Non-functional refactor of apply form components
